+++ .gitignore
@@ -1,14 +1,32 @@
-#added secrets.js file to gitignore to hide credentials
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# package directories
-node_modules
-jspm_packages
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+#added secrets.js file to gitignore to hide credentials
+secrets.js
 
 # Serverless directories
 .serverless
-
-.DS_Store
-
 # dotenv file, for api keys
 .env
-secrets.js
+
