+++ src/components/NavBar.scss
