+++ src/redux/reducers/index.js
@@ -39,7 +39,7 @@
 //   }
 // }
 
-import { INCREMENT, DECREMENT} from '../actions/index'
+import { INCREMENT, DECREMENT} from '../actions'
 
 
 const initial_state = {
