+++ src/views/public_dashboard/PublicDashboard.scss
